+++ sources/FEM/rf_pcs.h
@@ -693,7 +693,7 @@ public:
 			const int& ndx) const
 	{
 		double value = 0.;
-		for(int i=0; i<nodes.size(); ++i)
+		for(size_t i=0; i<nodes.size(); ++i)
 		{
 			value += GetNodeValue(nodes[i], ndx) * weights[i];
 		}
